+++ Camera/RealCamera.h
@@ -18,14 +18,14 @@ private:
     int _first_samples = 20;
 
     //Camera characteristics
-    double sensor_size = 3.5;
-    double focal_length = 5;
-    double lens_distance = 5.16748;
-    double aperture_size = 5;
-    double aperture_distance = 10;
-    double shutter_speed = 0.005;
+    float sensor_size = 3.5;
+    float focal_length = 5;
+    float lens_distance = 5.16748;
+    float aperture_size = 5;
+    float aperture_distance = 10;
+    float shutter_speed = 0.005;
 
-    static int additional_samples(double k) {
+    static int additional_samples(float k) {
         return (int) k + (int) (100.0 * sqrt(k));
     }
 
@@ -36,21 +36,21 @@ public:
     RealCamera(const ReferenceFrame &referenceFrame, const Scene &scene);
     RealCamera(const Vector3 &position, const Vector3 &point, Scene scene);
 
-    RealCamera(const ReferenceFrame &referenceFrame, const Scene &scene, double sensor_size, double focal_length,
-               double lens_distance, double aperture_size, double aperture_distance);
+    RealCamera(const ReferenceFrame &referenceFrame, const Scene &scene, float sensor_size, float focal_length,
+               float lens_distance, float aperture_size, float aperture_distance);
 
-    void set_sensor_size(double sensor_size);
-    void set_focal_length(double focal_length);
-    void set_lens_distance(double lens_distance);
-    void set_aperture_size(double aperture_size);
-    void set_aperture_distance(double aperture_distance);
-    void set_shutter_speed(double shutter_speed);
+    void set_sensor_size(float sensor_size);
+    void set_focal_length(float focal_length);
+    void set_lens_distance(float lens_distance);
+    void set_aperture_size(float aperture_size);
+    void set_aperture_distance(float aperture_distance);
+    void set_shutter_speed(float shutter_speed);
 
-    double get_focal_plane_distance() const;
-    double get_focal_plane_size() const;
+    float get_focal_plane_distance() const;
+    float get_focal_plane_size() const;
 
-    Picture picture(int height, int width, double time) override;
-    Color cast_ray(Vector2 position, double time);
+    Picture picture(int height, int width, float time) override;
+    Color cast_ray(Vector2 position, float time);
 
 };
