+++ Object/Object.cpp
@@ -11,13 +11,13 @@ void Object::add(Surface *surface) {
     furthest_distance = std::max(furthest_distance, surface->furthest_distance(0));
 }
 
-std::vector<std::pair<double, Surface *>> Object::intersections(const Line& ray, double time) {
+std::vector<std::pair<float, Surface *>> Object::intersections(const Line& ray, float time) {
 
-    std::vector<std::pair<double, Surface *>> result;
+    std::vector<std::pair<float, Surface *>> result;
 
     if (ray.distance(get_reference_frame().get_location(time), time) < furthest_distance)
         for (Surface * surface : surfaces) {
-            std::vector<std::pair<double, Surface *>> v = surface->intersections(get_reference_frame().to_ref_frame(ray, time), time);
+            std::vector<std::pair<float, Surface *>> v = surface->intersections(get_reference_frame().to_ref_frame(ray, time), time);
             result.insert(result.end(), v.begin(), v.end());
         }
