+++ Surface/Square.cpp
@@ -4,7 +4,7 @@
 
 #include "Square.h"
 
-Square::Square(const ReferenceFrame &referenceFrame, const Aspect &aspect, double side) : Surface(referenceFrame,
+Square::Square(const ReferenceFrame &referenceFrame, const Aspect &aspect, float side) : Surface(referenceFrame,
                                                                                                   aspect), side(side) {
     add(Vector2(side/2, side/2));
     add(Vector2(-side/2, side/2));
