+++ Utils/TimeGeometry/TimeUnitVector3.h
@@ -20,11 +20,11 @@ public:
     explicit TimeUnitVector3(const Vector3& v);
     explicit TimeUnitVector3(TimeVector3 v);
     TimeUnitVector3(TimeFunction x, TimeFunction y, TimeFunction z);
-    TimeUnitVector3(double x, double y, double z);
+    TimeUnitVector3(float x, float y, float z);
     explicit TimeUnitVector3(const UnitVector3& v);
     TimeUnitVector3();
 
-    [[nodiscard]] UnitVector3 time(double time) const;
+    [[nodiscard]] UnitVector3 time(float time) const;
 
 };
